+++ CHANGES.rst
@@ -1,11 +1,13 @@
-3.0.0 (2017-07-04)
+3.0.0 (2017-07-05)
 ==================
 
 base.py
 -------
 - Added the `ScreenShot` class containing data for a given screen shot (support the Numpy array interface [`ScreenShot.__array_interface__`])
+- Add `shot()` method to `MSSBase`. It takes the same arguments as the `save()` method.
 - Renamed `get_pixels` to `grab`. It now returns a `ScreenShot` object.
 - Moved `to_png` method to `tools.py`. It is now a simple function.
+- Changed `mon` default value of the `save()` method from `0` to `1`.
 - Removed `enum_display_monitors()` method. Use `monitors` property instead.
 - Removed `monitors` attribute. Use `monitors` property instead.
 - Removed `width` attribute. Use `ScreenShot.size[0]` attribute or `ScreenShot.width` property instead.